string: ARMv8.5 BTI: Add BTI support to assembly files.
This change addds the landing pads to the start of functions
implemented in assembly, by adding it to the ENTRY macro. To avoid
skipping it when using an alias, every ENTRY_ALIAS use must precede
the corresponding ENTRY.

Furthermore, the GNU property note is added to the assembly
files. Since none of the functions save LR to stack, both BTI and PAC
support are indicated.

Paddings before __strncmp_aarch64 and __strnlen_aarch64 were adjusted.

Change-Id: Iabae89c9f0b825d3afe493d285719c4621ab7c44