string: add a setting to disable GNU Property Notes
GNU Property Notes are only supported in recent tooling and older
tools may warn about them, so it makes sense to remove these notes
on a system where BTI is not supported anyway.

The actual BTI instructions should be kept in place to avoid
disturbing code layout.

-DWANT_GNU_PROPERTY=0 removes the .note.gnu.property section
from assembly files (ideally it would be based on the compiler
default setting, but there is no feature test macro for BTI and
PAC-RET).

Change-Id: Ia2647348547a9cb579b24adc2dd843604fb6cc04