string: Add MTE support to string tests.
Set taggs for every test case so that boundaries are as narrow as
possible. There is no handling of tag faults, so the test will
crash if there is a MTE problem.

The implementations that are not compatible are excluded, including
the standard symbols that may come from an mte incompatible libc.

Change-Id: I879956edae25401f16693d0a1f30376d6c1b6c2e